group invitations

+++ backend/KarmaCounterServer/KarmaCounterServer/Exceptions/PaymentNeededException.cs
@@ -6,6 +6,6 @@ namespace KarmaCounterServer.Exceptions
 {
     public class PaymentNeededException : ResponseException
     {
-        public PaymentNeededException(int amount) : base("Payment required", JsonConvert.DeserializeObject("{ \"amount\" : " + amount + " }"), HttpStatusCode.PaymentRequired) { }
+        public PaymentNeededException(double amount) : base("Payment required", JsonConvert.DeserializeObject("{ \"amount\" : " + amount + " }"), HttpStatusCode.PaymentRequired) { }
     }
 }
